fix(hostname): strip whitespace from hostname
There was an issue where in some cases the hostname would have a newline
at the end of it and in some cases it wouldn't. This adds code so any
whitespace at the end of the hostname is stripped - but only if you have
it automatically determine the hostname. If you're specifying the
hostname it will continue to use whatever you want.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>